+++ tests/testthat/test-filter-dm.R
@@ -1,4 +1,7 @@
 test_that("get_all_filtered_connected() calculates the paths correctly", {
+  # Only need to run for local sources
+  skip_if_remote_src()
+
   fc <-
     dm_more_complex() %>%
     dm_filter(tf_2, TRUE) %>%
@@ -55,6 +58,16 @@ test_that("get_all_filtered_connected() calculates the paths correctly", {
   expect_dm_error(
     dm_for_filter_w_cycle() %>% dm_filter(tf_1, a > 3) %>% dm_get_filtered_table("tf_3"),
     "no_cycles"
+  )
+
+  # Cycles in other components don't affect filtering
+  expect_equivalent_dm(
+    dm_for_filter_w_cycle() %>%
+      dm_add_tbl(tf_8 = tibble(r = 1)) %>%
+      dm_filter(tf_8, TRUE) %>%
+      dm_apply_filters(),
+    dm_for_filter_w_cycle() %>%
+      dm_add_tbl(tf_8 = tibble(r = 1))
   )
 
   # FIXME: fails, when it could actually work (check diagram of `dm_for_filter_w_cycle()`)